Consistent use of `sourceItem` as a variable name
In `_buildIndexFromSource` (previously called `_prepSource`) we
smartly called an item from the source `sourceItem`. I applied that
terminology to `_prepItem()` and `add()` to hopefully make it crystal
clear when you're dealing with an un-prepped item that is straight
from the source.

Maybe later it would make sense to rename `Indexed()` to
`PreparedItem()`.

Maybe this is too verbose, but I think it makes sense.